+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import DisplayArticles from "./components/DisplayArticles";
 import SingleArticle from "./components/SingleArticle";
 import { Switch, Route } from "react-router-dom";
-import MainHeader from "./components/MainHeader";
+import Header from "./components/Header";
 import LoginForm from "./components/LoginForm";
 import ProtectedRoute from "./components/ProtectedRoute";
 import BecomeSubscriber from "./components/BecomeSubscriber";
@@ -11,7 +11,7 @@ import RegistrationForm from "./components/RegistrationForm";
 const App = () => {
   return (
     <>
-      <MainHeader />
+      <Header />
       <Switch>
         <Route exact path="/category/:category" component={DisplayArticles} />
         <Route exact path="/articles/:id" component={SingleArticle} />
