+++ src/components/SingleArticle.jsx
@@ -1,4 +1,4 @@
-import { Image, Grid, Header, Button, Segment, Container } from "semantic-ui-react";
+import { Image, Grid, Header, Button, Segment } from "semantic-ui-react";
 import React, { useState, useEffect } from "react";
 import { useParams, Link } from "react-router-dom";
 import Articles from "../modules/articles";
@@ -29,8 +29,7 @@ const SingleArticle = () => {
       {message ? (
         <p data-cy="error-message">{message}</p>
       ) : (
-        <Container>
-        <Grid >
+        <Grid celled>
           <Grid.Row>
             <Grid.Column width={13}>
               <Header data-cy="title">{article.title}</Header>
@@ -54,13 +53,12 @@ const SingleArticle = () => {
 
             <Grid.Column width={5}></Grid.Column>
           </Grid.Row>
-          { article.premium && currentUser.role !== "subscriber" && (
+          {article.premium && currentUser.role !== "subscriber" && (
             <Segment color="black" textAlign="center">
               <h4 data-cy="premium-message">
                 This is a premium article, become a subscriber to read full
                 content{" "}
               </h4>
-              
               <Button
                 as={Link}
                 to={authenticated ? "/become-subscriber" : "/login"}
@@ -74,7 +72,6 @@ const SingleArticle = () => {
             </Segment>
           )}
         </Grid>
-        </Container>
       )}
     </>
   );
