+++ src/state/reducers/rootReducer.js
@@ -6,7 +6,6 @@ const rootReducer = (state = initialState, action) => {
       return {
         ...state,
         ...action.payload,
-        role: action.payload.role,
       };
     case "USER_IS_SUBSCRIBER":
       return {
